Quirky but worky.
Basically works OK.  A few issues.

1) It doesn't appear possible (or easy at least) to cause the host computer
to re-read the disk after it is inserted.  This makes development a bit of a pain...

Probably possibly to have it disconnect in software, but not sure.

Anyway - this firmware seems to track that on Car F very closely.  Need some testing
and breaking, but should be close.

Close #9
Close #5
Close #12